+++ src/components/lights/LightBase.ts
@@ -99,11 +99,20 @@ export class LightBase extends ComponentBase implements ILight {
     public onEnable(): void {
         this.onChange();
         EntityCollect.instance.addLight(this.transform.scene3D, this);
+
+        if (this._castShadow) {
+            this.needUpdateShadow = true;
+            ShadowLightsCollect.addShadowLight(this);
+        }
     }
 
     public onDisable(): void {
         this.onChange();
         EntityCollect.instance.removeLight(this.transform.scene3D, this);
+
+        if (this._castShadow) {
+            ShadowLightsCollect.removeShadowLight(this);
+        }
     }
 
     public set iesProfiles(iesProfiles: IESProfiles) {
@@ -241,11 +250,11 @@ export class LightBase extends ComponentBase implements ILight {
         return this.lightData.direction;
     }
 
-    public destroy(): void {
+    public destroy(force?: boolean): void {
         this.bindOnChange = null;
         this.transform.eventDispatcher.removeEventListener(Transform.ROTATION_ONCHANGE, this.onRotChange, this);
         this.transform.eventDispatcher.removeEventListener(Transform.SCALE_ONCHANGE, this.onScaleChange, this);
-        super.destroy();
+        super.destroy(force);
     }
 
 }