+++ test/components/PostEffects_HDRBloom.test.ts
@@ -2,7 +2,9 @@ import { test, end } from '../util'
 import { Camera3D, CameraUtil, Engine3D, HDRBloomPost, Object3D, PostProcessingComponent, Scene3D, View3D } from '@orillusion/core';
 
 await test('Post HDRBloomPost test', async () => {
-    let suc = await Engine3D.init();
+    await Engine3D.init();
+    Engine3D.frameRate = 10;
+
     let view = new View3D();
     view.scene = new Scene3D();
     view.camera = CameraUtil.createCamera3DObject(view.scene, "camera");